+++ src/Traits/HasAllTraits.php
@@ -2,13 +2,14 @@
 
 namespace Rappasoft\LaravelLivewireTables\Traits;
 
-use Rappasoft\LaravelLivewireTables\Traits\Core\HasCustomAttributes;
+use Rappasoft\LaravelLivewireTables\Traits\Core\{HasCustomAttributes,HasLocalisations};
 use Rappasoft\LaravelLivewireTables\Views\Traits\Core\HasTheme;
 
 trait HasAllTraits
 {
     // Note Specific Order Below!
     use WithTableHooks;
+    use HasLocalisations;
     use WithLoadingPlaceholder;
     use HasTheme;
     use ComponentUtilities,