+++ src/Traits/Helpers/ComponentHelpers.php
@@ -63,7 +63,7 @@ trait ComponentHelpers
     public function getEmptyMessage(): string
     {
         if ($this->emptyMessage == 'No items found, try to broaden your search') {
-            return __('livewire-tables::core.No items found, try to broaden your search');
+            return __($this->getLocalisationPath().'No items found, try to broaden your search');
         }
 
         return $this->emptyMessage;