+++ src/Views/Traits/Core/HasView.php
@@ -3,7 +3,6 @@
 namespace Rappasoft\LaravelLivewireTables\Views\Traits\Core;
 
 use Rappasoft\LaravelLivewireTables\Exceptions\DataTableConfigurationException;
-use Rappasoft\LaravelLivewireTables\Views\{Column,Filter};
 
 trait HasView
 {