+++ src/Views/Traits/IsFilter.php
@@ -2,14 +2,16 @@
 
 namespace Rappasoft\LaravelLivewireTables\Views\Traits;
 
+use Rappasoft\LaravelLivewireTables\Traits\Core\HasLocalisations;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Configuration\FilterConfiguration;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Core\{HasConfig,HasLabelAttributes,HasView};
-use Rappasoft\LaravelLivewireTables\Views\Traits\Filters\{HasCustomPosition,HasVisibility};
+use Rappasoft\LaravelLivewireTables\Views\Traits\Filters\{HasCustomPosition, HasVisibility};
 use Rappasoft\LaravelLivewireTables\Views\Traits\Helpers\FilterHelpers;
 
 trait IsFilter
 {
-    use FilterConfiguration,
+    use HasLocalisations,
+        FilterConfiguration,
         FilterHelpers,
         HasConfig,
         HasCustomPosition,